Support to group elements in container records
Records like 'BOF/EOF' and 'BEGIN/END' are now
grouped together in containers. We pick the last
sibling of a container record and name the container
as a container of it's sibling record. For e.g. if
a trail of BEGIN/END follows an OBJ Record, the trail
is wrapped in a container with name OBJ Records